Shrine path/gatherloot changes (#7)
* Shrine gatherloot path update

middle ? platform causes bad movement, this path avoids that. needs a bit of testing with movespeeds and just that it gets all the loot

* Remove unnecessarycamera movement + unnecessary if MoveMethod

Shrine path needs to be tested with abnormal movespeeds

* remove send {raw}

messed up gather pattern

It *should* be mostly stable now. but i will continue testing

* Update path timing

Reduces the chances of jumping early

* Make last jump to shrine more consistent